# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "Example info",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and generate Bearer token",
                "description": "Handle an incoming authentication request.",
                "operationId": "44212a9096e4b09358281e9ec8a0701d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User's email",
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSchema"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Handle an incoming registration request.",
                "operationId": "a718f172ff3ac464d723835815f8fb57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User's name",
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "email": {
                                        "description": "User's email",
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "password": {
                                        "description": "User's password",
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "User created successfully",
                                            "type": "string",
                                            "example": "User created successfully",
                                            "nullable": "false"
                                        },
                                        "token": {
                                            "description": "Token",
                                            "type": "string",
                                            "nullable": "false"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Clears user tokens",
                "description": "Handle an incoming logout request.",
                "operationId": "79854151e609bd20032aa5aa4d5a35c0",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                }
            }
        },
        "/api/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get list of notes",
                "description": "Returns list of notes",
                "operationId": "getNotesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NoteSchema"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Store new note",
                "description": "Returns note data",
                "operationId": "storeNote",
                "parameters": [
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Note text",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "description": "Note text",
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/notes/{id}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get note information",
                "description": "Returns note data",
                "operationId": "getNoteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Notes"
                ],
                "summary": "Update existing note",
                "description": "Returns updated note data",
                "operationId": "updateNote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "description": "Note text",
                                        "type": "string",
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Delete existing note",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteNote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserSchema": {
                "properties": {
                    "id": {
                        "description": "User identifier",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "admin",
                        "nullable": "false"
                    },
                    "email": {
                        "description": "User E-mail",
                        "type": "string",
                        "example": "kellen.boyer@example.com",
                        "nullable": "false"
                    },
                    "role": {
                        "description": "User role",
                        "type": "string",
                        "example": "user",
                        "nullable": "false"
                    },
                    "updated_at": {
                        "description": "User updated date",
                        "type": "string",
                        "example": "2023-11-15T05:34:17.000000Z",
                        "nullable": "false"
                    },
                    "created_at": {
                        "description": "User created date",
                        "type": "string",
                        "example": "2023-11-15T05:34:17.000000Z",
                        "nullable": "false"
                    }
                },
                "type": "object"
            },
            "NoteSchema": {
                "properties": {
                    "id": {
                        "description": "Note identifier",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "text": {
                        "description": "Note text",
                        "type": "string",
                        "example": "Note text",
                        "nullable": "false"
                    },
                    "user_id": {
                        "description": "User E-mail",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "updated_at": {
                        "description": "Note updated date",
                        "type": "string",
                        "example": "2023-11-15T05:34:17.000000Z",
                        "nullable": "false"
                    },
                    "created_at": {
                        "description": "Note created date",
                        "type": "string",
                        "example": "2023-11-15T05:34:17.000000Z",
                        "nullable": "false"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Autetinification API"
        },
        {
            "name": "Notes",
            "description": "API Endpoints of Notes"
        }
    ]
}